Factor out Kore.AST.Identifier and Kore.Sort
