+++ src/main/haskell/kore/src/Kore/ASTVerifier/ModuleVerifier.hs
@@ -16,7 +16,7 @@ import qualified Data.Map as Map
 import           Data.Text
                  ( Text )
 
-import           Kore.AST.Common
+import           Kore.AST.Kore
 import           Kore.AST.Sentence
 import           Kore.ASTVerifier.AttributesVerifier
 import           Kore.ASTVerifier.Error
