+++ src/main/haskell/kore/src/Kore/Proof/Util.hs
@@ -55,8 +55,7 @@ import           Data.Text
                  ( Text )
 import qualified Data.Text as Text
 
-import Kore.AST.Common
-import Kore.AST.MetaOrObject
+import Kore.AST.Pure
 import Kore.ASTUtils.SmartConstructors
 import Kore.ASTUtils.SmartPatterns
 import Kore.IndexedModule.MetadataTools
