+++ src/main/haskell/kore/src/Kore/Step/Simplification/AndTerms.hs
@@ -33,9 +33,7 @@ import qualified Data.Set as Set
 import           Prelude hiding
                  ( concat )
 
-import           Kore.AST.Common
-                 ( Sort, SortedVariable, SymbolOrAlias (..) )
-import           Kore.AST.MetaOrObject
+import           Kore.AST.Pure
 import           Kore.ASTUtils.SmartConstructors
                  ( mkAnd, mkApp, mkBottom, mkTop )
 import           Kore.ASTUtils.SmartPatterns
