+++ src/main/haskell/kore/test/Test/Kore/ASTVerifier/DefinitionVerifier/UniqueNames.hs
@@ -7,8 +7,7 @@ import Test.Tasty
 
 import qualified Data.Text as Text
 
-import Kore.AST.Common
-import Kore.AST.MetaOrObject
+import Kore.AST.Kore
 import Kore.AST.Sentence
 import Kore.Error
 import Kore.Implicit.ImplicitSorts
