+++ src/main/haskell/kore/test/Test/Kore/Builtin/List.hs
@@ -12,8 +12,7 @@ import           Data.Sequence
                  ( Seq )
 import qualified Data.Sequence as Seq
 
-import           Kore.AST.Common
-import           Kore.AST.MetaOrObject
+import           Kore.AST.Pure
 import qualified Kore.ASTUtils.SmartConstructors as Kore
 import           Kore.ASTUtils.SmartPatterns
 import qualified Kore.Builtin.List as List
