+++ src/main/haskell/kore/test/Test/Kore/Step/Condition/Evaluator.hs
@@ -14,8 +14,7 @@ import           Data.Text
                  ( Text )
 import qualified Data.Text as Text
 
-import           Kore.AST.Common
-import           Kore.AST.MetaOrObject
+import           Kore.AST.Pure
 import           Kore.ASTUtils.SmartConstructors
 import           Kore.ASTUtils.SmartPatterns
 import           Kore.IndexedModule.MetadataTools
