+++ src/main/haskell/kore/test/Test/Kore/Step/ExpandedPattern.hs
@@ -9,10 +9,7 @@ import Test.Tasty
 import Test.Tasty.HUnit
        ( assertEqual, testCase )
 
-import           Kore.AST.Common
-                 ( AstLocation (..), Id (..), Sort (..), SortVariable (..),
-                 SortedVariable (..) )
-import           Kore.AST.MetaOrObject
+import           Kore.AST.Pure
 import           Kore.ASTHelpers
                  ( ApplicationSorts (..) )
 import           Kore.ASTUtils.SmartConstructors
