+++ src/main/haskell/kore/test/Test/Kore/Step/Simplification/Application.hs
@@ -11,8 +11,7 @@ import           Data.Reflection
 import           Data.These
                  ( These (That) )
 
-import           Kore.AST.Common
-import           Kore.AST.MetaOrObject
+import           Kore.AST.Pure
 import           Kore.ASTHelpers
                  ( ApplicationSorts (..) )
 import           Kore.ASTUtils.SmartConstructors
